Re-Enable carthage support
Also update for swift 4
