+++ LFLoginController/LFLoginController.h
@@ -1,0 +1,19 @@
+//
+//  LFLoginController.h
+//  LFLoginController
+//
+//  Created by Rachel Brindle on 2/23/18.
+//  Copyright © 2018 Lucas Farah. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for LFLoginController.
+FOUNDATION_EXPORT double LFLoginControllerVersionNumber;
+
+//! Project version string for LFLoginController.
+FOUNDATION_EXPORT const unsigned char LFLoginControllerVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <LFLoginController/PublicHeader.h>
+
+
