+++ LFLoginControllerExample/AppDelegate.swift
