+++ LFLoginControllerExample/Assets.xcassets/AwesomeLabsLogoWhite.imageset/AwesomeLabsLogoWhite.png
