+++ LFLoginControllerExample/Assets.xcassets/AwesomeLabsLogoWhite.imageset/Contents.json
