+++ LFLoginControllerExample/Assets.xcassets/Contents.json
