+++ LFLoginControllerExample/ViewController.swift
@@ -9,6 +9,7 @@
 // swiftlint:disable trailing_whitespace
 
 import UIKit
+import LFLoginController
 
 class ViewController: UIViewController {
     
@@ -31,11 +32,6 @@ class ViewController: UIViewController {
     @IBAction func butLoginTapped(sender: AnyObject) {
         
         self.navigationController?.pushViewController(controller, animated: true)
-    }
-    
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
     }
 }
