+++ LFLoginControllerUITests/Info.plist
