+++ Sources/AutoCompleteTextField.swift
@@ -207,7 +207,7 @@ open class AutoCompleteTextField: UITextField {
         }else{
 
             let suggestedString = suggestedStrings.sorted(by: { (elementOne, elementTwo) -> Bool in
-                return elementOne.characters.count < elementTwo.characters.count
+                return elementOne.count < elementTwo.count
             }).first ?? ""
             return performStringReplacement(suggestedString, stringFilter: stringFilter)
         }
@@ -233,7 +233,7 @@ open class AutoCompleteTextField: UITextField {
         let paragraphStyle = NSMutableParagraphStyle()
         paragraphStyle.lineBreakMode = .byCharWrapping
         
-        let textAttributes: [String: AnyObject] = [NSFontAttributeName: font!, NSParagraphStyleAttributeName: paragraphStyle]
+        let textAttributes: [NSAttributedStringKey: AnyObject] = [NSAttributedStringKey.font: font!, NSAttributedStringKey.paragraphStyle: paragraphStyle]
         
         let drawingOptions: NSStringDrawingOptions = [.usesLineFragmentOrigin, .usesFontLeading]
         
@@ -299,14 +299,14 @@ open class AutoCompleteTextField: UITextField {
     
     // MARK: - Internal Controls
     
-    internal func autoCompleteButtonDidTapped(_ sender: UIButton) {
+    @objc internal func autoCompleteButtonDidTapped(_ sender: UIButton) {
         endEditing(true)
         
         processAutoCompleteEvent()
         commitAutocompleteText()
     }
     
-    internal func autoCompleteTextFieldDidChanged(_ textField: UITextField) {
+    @objc internal func autoCompleteTextFieldDidChanged(_ textField: UITextField) {
         
         processAutoCompleteEvent()
     }
